+++ apps/pioneer-lib/src/app/components/PathWizard/index.tsx
@@ -1,17 +1,16 @@
 import React, { useState, useEffect } from 'react';
 import {
-  Box, Button, Flex, FormControl, FormLabel, Input, Stack, useToast, Avatar, Heading,
-  Menu, MenuButton, MenuList, MenuItem, HStack, Icon, Text, useDisclosure, Select, Tooltip, Collapse, Textarea
+  Box, Button, Flex, FormControl, FormLabel, Input, useToast, Avatar, Menu, MenuButton, MenuList, MenuItem, HStack, Text
 } from '@chakra-ui/react';
-import { COIN_MAP_LONG, getPaths, addressNListToBIP32, bip32ToAddressNList } from '@pioneer-platform/pioneer-coins';
-import { ChevronDownIcon, ChevronLeftIcon, ChevronRightIcon, InfoOutlineIcon, ChevronUpIcon } from '@chakra-ui/icons';
-// import { usePioneer } from '@coinmasters/pioneer-react';
+import { ChevronDownIcon, ChevronLeftIcon, ChevronRightIcon } from '@chakra-ui/icons';
+import { COIN_MAP_LONG, getPaths } from '@pioneer-platform/pioneer-coins';
 import { availableChainsByWallet, ChainToNetworkId, getChainEnumValue, NetworkIdToChain } from '@coinmasters/types';
 import Path from '../../components/Path';
+import StepTwo from './StepTwo';
 
 const StepIndicator = ({ steps, currentStep }:any) => (
   <HStack spacing={4} justify="center" mb={6}>
-    {steps.map((step: any, index: any) => (
+    {steps.map((step:any, index:any) => (
       <Box
         key={index}
         p={2}
@@ -30,92 +29,11 @@ const StepIndicator = ({ steps, currentStep }:any) => (
   </HStack>
 );
 
-const BIP32Description = () => (
-  <Box mb={6}>
-    <Text mb={4}>A BIP32 path specifies a unique hierarchy for deriving cryptographic keys from a master seed. Each segment of the path has a specific purpose:</Text>
-    {/*<Text><b>purpose</b> - Indicates the high-level purpose of this key hierarchy, often tied to a standard like BIP44.</Text>*/}
-    {/*<Text><b>coin_type</b> - Identifies the cryptocurrency to which the keys belong, ensuring separation between chains.</Text>*/}
-    {/*<Text><b>account</b> - Separates keys into logical groupings that can represent accounts.</Text>*/}
-    {/*<Text><b>change</b> - Differentiates between keys used for receiving transactions and keys used for change in transactions.</Text>*/}
-    {/*<Text><b>address_index</b> - Identifies individual addresses within the account and change categories.</Text>*/}
-  </Box>
-);
-
-const StepTwo = ({ entry, onEntryChange }: any) => {
-  const [type, setType] = useState(entry.type);
-  const { isOpen, onToggle } = useDisclosure();
-  const [isPathValid, setIsPathValid] = useState(true);
-
-  // Assuming bip32ToAddressNList is imported or defined elsewhere
-  const handlePathChange = (e:any) => {
-    const newPath = e.target.value;
-    try {
-      // Assuming bip32ToAddressNList converts a BIP32 path to an addressNList format
-      const newAddressNList = bip32ToAddressNList(newPath);
-
-      // Update both addressNList and addressNListMaster based on the converted value
-      // For addressNListMaster, append the additional segments (0, 0) for the master key derivation
-      onEntryChange({
-        ...entry,
-        addressNList: newAddressNList,
-        addressNListMaster: [...newAddressNList, 0, 0], // Appends 0'/0' at the end for the master
-        path: newPath
-      });
-      setIsPathValid(true); // Path is valid
-    } catch (error) {
-      console.error("Error converting BIP32 path to addressNList:", error);
-      setIsPathValid(false); // Path is invalid but keep it in the form for user to edit
-    }
-  };
-
-  return (
-    <Box>
-      {/*<BIP32Description />*/}
-      <FormControl isInvalid={!isPathValid}>
-        <FormLabel>Primary BIP32 Path</FormLabel>
-        <Input
-          name="path"
-          // Safely handle the conversion with a check or default value
-          value={entry.path || (entry.addressNList ? addressNListToBIP32(entry.addressNList) : '')}
-          onChange={handlePathChange}
-          placeholder="m/44'/0'/0'/0/0"
-          borderColor={isPathValid ? "green.500" : undefined} // Use borderColor for valid paths
-        />
-        {!isPathValid && (
-          <Text color="red.500" mt={2}>
-            Not a valid BIP32 path. Please correct the path.
-          </Text>
-        )}
-      </FormControl>
-
-      <Flex alignItems="center" mt={4}>
-        <Text flexShrink={0}>Advanced</Text>
-        <ChevronDownIcon
-          onClick={onToggle}
-          // icon={isOpen ? <ChevronUpIcon /> : <ChevronDownIcon />}
-          aria-label="Toggle advanced JSON edit"
-          ml={2}
-        />
-      </Flex>
-
-      <Collapse in={isOpen} animateOpacity>
-        <Box mt={4} p={4} border="1px" borderColor="gray.300">
-          <Textarea
-            value={JSON.stringify(entry, null, 2)}
-            onChange={(e) => onEntryChange(JSON.parse(e.target.value))}
-            h="200px"
-          />
-        </Box>
-      </Collapse>
-    </Box>
-  );
-};
-
-const PathWizard = () => {
+const PathWizard = ({ usePioneer, networkId }:any) => {
   const toast = useToast();
-  const [currentStep, setCurrentStep] = useState(1);
+  const [currentStep, setCurrentStep] = useState(networkId ? 2 : 1);
   const [allChains, setAllChains] = useState([]);
-  const [selectedChain, setSelectedChain] = useState('');
+  const [selectedChain, setSelectedChain] = useState(networkId || '');
   const [path, setPath] = useState<any>({});
   const [pathName, setPathName] = useState('');
 
@@ -142,24 +60,21 @@ const PathWizard = () => {
     console.log("path: ", path);
   }, [path]);
 
-  let getPathForChain = async function(){
-    try{
-      let defaultPath = getPaths([selectedChain])
+  const getPathForChain = async () => {
+    try {
+      let defaultPath = getPaths([selectedChain]);
       console.log('defaultPath: ', defaultPath[0]);
-      setPath(defaultPath[0]); // Update your state accordingly
-      //
-
-    }catch(e){
-      console.error(e)
+      setPath(defaultPath[0]);
+    } catch (e) {
+      console.error(e);
     }
-  }
+  };
 
   useEffect(() => {
     if (currentStep === 2) {
       getPathForChain();
     }
-  }, [currentStep, selectedChain]); // Ensure this effect depends on both currentStep and selectedChain
-
+  }, [currentStep, selectedChain]);
 
   const nextStep = () => setCurrentStep(currentStep < 3 ? currentStep + 1 : currentStep);
   const prevStep = () => setCurrentStep(currentStep > 1 ? currentStep - 1 : currentStep);
@@ -168,26 +83,19 @@ const PathWizard = () => {
     if (typeof window !== 'undefined') {
       const lastConnectedWallet = window.localStorage.getItem('lastConnectedWallet');
       if (!lastConnectedWallet) {
-        // Handle the case where there is no last connected wallet, if necessary
         return;
       }
 
       const walletType = lastConnectedWallet.split(':')[0];
-
-      // Retrieve the existing paths array from localStorage, or initialize it as an empty array if not present
       const existingPathsStr = window.localStorage.getItem(walletType + ':paths:add');
       const existingPaths = existingPathsStr ? JSON.parse(existingPathsStr) : [];
 
-      // Check if a path with the same addressNList already exists
-      const isDuplicate = existingPaths.some((existingPath: any) =>
+      const isDuplicate = existingPaths.some((existingPath:any) =>
         JSON.stringify(existingPath.addressNList) === JSON.stringify(path.addressNList)
       );
 
       if (!isDuplicate) {
-        // Push the new path into the array if it's not a duplicate
         existingPaths.push(path);
-
-        // Save the updated array back to localStorage
         window.localStorage.setItem(walletType + ':paths:add', JSON.stringify(existingPaths));
 
         toast({
@@ -199,7 +107,6 @@ const PathWizard = () => {
         });
         window.location.reload();
       } else {
-        // Optionally notify the user that the path was not added because it's a duplicate
         toast({
           title: "Duplicate Path",
           description: "This path already exists and was not added again.",
@@ -228,7 +135,7 @@ const PathWizard = () => {
                   <MenuItem key={chain} minH="40px" onClick={() => setSelectedChain(chain)}>
                     <Avatar
                       size="xs"
-                      src={`https://pioneers.dev/coins/${COIN_MAP_LONG[NetworkIdToChain[chain] as keyof typeof COIN_MAP_LONG]}.png`} // Modify this line according to your mapping of chains to images
+                      src={`https://pioneers.dev/coins/${COIN_MAP_LONG[NetworkIdToChain[chain as keyof typeof NetworkIdToChain] as keyof typeof COIN_MAP_LONG]}.png`}
                       mr={3}
                     />
                     <Text>{chain}</Text>
@@ -241,8 +148,10 @@ const PathWizard = () => {
 
         {currentStep === 2 && (
           <StepTwo
+            usePioneer={usePioneer}
+            networkId={networkId}
             entry={path}
-            onEntryChange={(updatedEntry: any) => setPath(updatedEntry)}
+            onEntryChange={(updatedEntry:any) => setPath(updatedEntry)}
           />
         )}
 
@@ -251,14 +160,11 @@ const PathWizard = () => {
             <FormLabel>Name Your Path</FormLabel>
             <Input placeholder="Path name" value={pathName} onChange={(e) => setPathName(e.target.value)} />
 
-            rewview:
-            <Path path={path} />
+            <Path usePioneer={usePioneer} path={path} />
 
             <Button mt={4} colorScheme="blue" onClick={handleSave}>Save</Button>
           </FormControl>
         )}
-
-
 
         <HStack spacing={4} mt={6}>
           <Button leftIcon={<ChevronLeftIcon />} onClick={prevStep} isDisabled={currentStep === 1}>