+++ packages/toolboxes/utxo/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @swapkit/toolbox-utxo
 
+## 11.0.40
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.7.40
+
+## 11.0.39
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.7.39
+
 ## 11.0.38
 
 ### Patch Changes