+++ packages/wallets/okx/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @coinmasters/wallet-okx
 
+## 11.0.40
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@11.0.40
+  - @coinmasters/toolbox-utxo@11.0.40
+  - @coinmasters/types@4.7.40
+  - @coinmasters/toolbox-evm@11.0.40
+
+## 11.0.39
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@11.0.39
+  - @coinmasters/toolbox-utxo@11.0.39
+  - @coinmasters/types@4.7.39
+  - @coinmasters/toolbox-evm@11.0.39
+
 ## 11.0.38
 
 ### Patch Changes